map finished except one bug
the map crashes when clicking twice and  zooming into a new chosen location, because of the onViewportChange event at the geocoder. This functionality disabled for now, and everything works very well without it. However, we  should come back to it later in case we have time in the end of the project